Add Is Published Boolean in Add and Edit Form in Cruise and Package Add Migrations to Add Column [is_published] Add Validation Delete Cruise/Package/Adventure for Checking Have Reservations or Not Add Query in Cruise and package for Published Data Fix Bug in New Bokking Confirmation Mail

+++ app/Http/Controllers/ApiV2/Admin/CruiseController.php
@@ -169,9 +169,14 @@ class CruiseController extends Controller
     
     public function destroy(Cruise $cruise)
     {
-        if( $cruise->delete() )
-            return response()->json(['message' =>'operation done successfully', 'status' => 200]);
-
+        $checkBookings = Booking::whereModelType('App\Models\Cruise')->whereModelId($cruise->id)->count();
+        if ($checkBookings > 0) {
+            return response()->json(['message' =>'you can not delete that because you have reservations', 'status' => 400]);
+        } else {
+            if($cruise->delete()) {
+                return response()->json(['message' =>'operation done successfully', 'status' => 200]);
+            }
+        }
         return response()->json(['message' =>'operation failed', 'status' => 400]);
     }
